Give line number column the same bg colour as the content
There seems to be one column of padding between the line number and the
content. This column, however shares the same background colour as the line
number.

I was not happy with the way that there was no gap between the the main
content and the line number column, so I decided to make the line number
columns's background the same as the content background colour.

I then "dimmed" the line number's foreground colour slightly so as to
provide a bit of visual separation between the numbers and the content.